Modified Joboffer DO/DTO
