--- JFSData/src/main/java/jfs/data/dataobjects/LocationDO.java
@@ -1,16 +1,0 @@
-package jfs.data.dataobjects;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * Created by Deepika on 12/3/2015.
- */
-public class LocationDO {
-    public List<String> coordindates = new ArrayList<>();
-    public String type = "Point";
-
-    public LocationDO(){
-
-    }
-}
